dicom: move error propagation, cleanup defs, VFS to decoder module
The synthetic driver will need access to the error propagation functions
and cleanup definitions.  It makes sense for the VFS callbacks to be in
the decoder too, so move those as well.
